(OSX) Some forward-compatibility build fixes

+++ apple/OSX/settings.m
@@ -23,6 +23,8 @@
 #include "../../input/input_common.h"
 #include "../../input/input_keymaps.h"
 
+#include <objc/runtime.h>
+
 static void* const associated_name_tag = (void*)&associated_name_tag;
 
 @interface RAInputBinder : NSWindow
@@ -209,12 +211,12 @@ NSWindowDelegate>
 }
 
 #pragma mark Section Table
-- (int32_t)numberOfRowsInTableView:(NSTableView*)view
+- (NSInteger)numberOfRowsInTableView:(NSTableView*)view
 {
    return self.settings.count;
 }
 
-- (id)tableView:(NSTableView *)tableView objectValueForTableColumn:(NSTableColumn *)tableColumn row:(int32_t)row
+- (id)tableView:(NSTableView *)tableView objectValueForTableColumn:(NSTableColumn *)tableColumn row:(NSInteger)row
 {
 #if defined(MAC_OS_X_VERSION_10_6)
    return objc_getAssociatedObject([self.settings objectAtIndex:row], associated_name_tag);
@@ -231,12 +233,12 @@ NSWindowDelegate>
 }
 
 #pragma mark Setting Outline
-- (int32_t)outlineView:(NSOutlineView *)outlineView numberOfChildrenOfItem:(id)item
+- (NSInteger)outlineView:(NSOutlineView *)outlineView numberOfChildrenOfItem:(id)item
 {
    return (item == nil) ? [self.currentGroup count] : [item count];
 }
 
-- (id)outlineView:(NSOutlineView *)outlineView child:(int32_t)idx ofItem:(id)item
+- (id)outlineView:(NSOutlineView *)outlineView child:(NSInteger)idx ofItem:(id)item
 {
    return (item == nil) ? [self.currentGroup objectAtIndex:idx] : [item objectAtIndex:idx];
 }
